refactor：重构服务层接口

+++ my-shop-web-admin/src/main/java/com/wenwl/my/shop/web/admin/service/impl/TbContentServiceImpl.java
@@ -88,7 +88,7 @@ public class TbContentServiceImpl implements TbContentService {
      * @return
      */
     @Override
-    public long batchDelete(String[] ids) {
+    public int batchDelete(String[] ids) {
         return tbContentDao.batchDelete(ids);
     }
