+++ my-shop-web-admin/src/main/java/com/wenwl/my/shop/web/admin/service/impl/TbUserServiceImpl.java
@@ -111,7 +111,7 @@ public class TbUserServiceImpl implements TbUserService {
      * @return
      */
     @Override
-    public long batchDelete(String[] ids) {
+    public int batchDelete(String[] ids) {
         return tbUserDao.batchDelete(ids);
     }
